Changed to packaged "library" with install script
install.sh moves lilysplit into a directory in $PATH, if possible, and
the include files are now in gomp/ but can, for now, still be included
via lilyparts/*.

Now, this repository needs not be linked as a submodule, but can be
checked out separately somewhere on the machine. Once ./install.sh has
been run successfully, lilysplit takes care of everything.