FIX/lib move bookOutputSuffix to score.ly
With this architecture, paper.ly isn't strictly necessary.

In 9f5cc94030e384b2fcfed7bebbc479ebefdf26be (v1.0), we moved it the
other way, citing problems with lilypond-book. I don't use that anymore,
so I will deal with it when the need arises. For now, I want paper.ly to
be optional.