FIX spans
Spans stopped working because \ottava 0 does no longer start an
ottavation bracket if none is there.

Therefore, we implement everything in scheme and set middleCPos to what
it hopefully was before the ottavation.

We need to be hopeful because I cannot figure out how to store the
previous middleCPosition in a variable that will also be read by
SimultaneousMusic (which we need for \divisi, but that's for later).
Instead, if we aren't on a G_8 clef, we need to pass the correct current
middleCPosition.

Hope alone doesn't cut it, though, so we test the middleCPosition and
throw an error if it isn't the one the caller promised it should be.